pyproject.toml
[project]
name = "property-tools"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas (>=2.3.2,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "geopy (>=2.4.1,<3.0.0)",
    "shapely (>=2.1.2,<3.0.0)",
    "folium (>=0.20.0,<0.21.0)",
    "googlemaps (>=4.10.0,<5.0.0)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "pdfplumber (>=0.11.7,<0.12.0)",
    "streamlit (>=1.50.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
